fixed loader

+++ includes/wood_woodpacker.h
@@ -25,7 +25,7 @@ typedef struct loaderData
     uint64_t exec_pt_load_size;
     uint64_t offset_to_old_entryoint;
     char key[MAX_KEY_SIZE];
-    uint8_t key_len;
+    uint64_t key_len;
 } loaderData;
 
 // loader.asm
